pyproject.toml
[project]
name = "industrial-rag-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "elasticsearch>=9.1.0",
    "fastapi>=0.116.1",
    "google-genai>=1.30.0",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.3",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
